+++ crates/pine-runtime/src/tests/strategy.rs
@@ -3606,8 +3606,11 @@ plot(strategy.closedtrades)
 plot(strategy.grossprofit)
 plot(strategy.grossloss)
 plot(strategy.avg_trade)
+plot(strategy.avg_trade_percent)
 plot(strategy.avg_winning_trade)
+plot(strategy.avg_winning_trade_percent)
 plot(strategy.avg_losing_trade)
+plot(strategy.avg_losing_trade_percent)
 "#,
     );
     let analysis = analyze_source(&source);
@@ -3736,6 +3739,20 @@ plot(strategy.avg_losing_trade)
         vec![
             PineValue::Na,
             PineValue::Na,
+            PineValue::Float(50.0),
+            PineValue::Float(50.0),
+            PineValue::Float(50.0),
+            PineValue::Float(0.0),
+            PineValue::Float(0.0),
+            PineValue::Float(0.0),
+            PineValue::Float(0.0),
+        ]
+    );
+    assert_eq!(
+        result.plots[8].values,
+        vec![
+            PineValue::Na,
+            PineValue::Na,
             PineValue::Float(1.0),
             PineValue::Float(1.0),
             PineValue::Float(1.0),
@@ -3746,7 +3763,21 @@ plot(strategy.avg_losing_trade)
         ]
     );
     assert_eq!(
-        result.plots[8].values,
+        result.plots[9].values,
+        vec![
+            PineValue::Na,
+            PineValue::Na,
+            PineValue::Float(50.0),
+            PineValue::Float(50.0),
+            PineValue::Float(50.0),
+            PineValue::Float(50.0),
+            PineValue::Float(50.0),
+            PineValue::Float(50.0),
+            PineValue::Float(50.0),
+        ]
+    );
+    assert_eq!(
+        result.plots[10].values,
         vec![
             PineValue::Na,
             PineValue::Na,
@@ -3757,6 +3788,20 @@ plot(strategy.avg_losing_trade)
             PineValue::Float(2.0),
             PineValue::Float(2.0),
             PineValue::Float(2.0),
+        ]
+    );
+    assert_eq!(
+        result.plots[11].values,
+        vec![
+            PineValue::Na,
+            PineValue::Na,
+            PineValue::Na,
+            PineValue::Na,
+            PineValue::Na,
+            PineValue::Float(50.0),
+            PineValue::Float(50.0),
+            PineValue::Float(50.0),
+            PineValue::Float(50.0),
         ]
     );
     let strategy = result.strategy.as_ref().expect("strategy result");